gen_cfile.py: fix null character size counting
The change from Python2 to Python 3 support introduced a bug where strings in
.od files containing the NULL character "\x00" are no longer interpreted as a
byte of size one but as a string of size three. This bug is because Python2
treats strings as byte arrays, and Python3 treats strings as immutable
sequences of Unicode characters.

The fix is relatively simple: Check in a few places in gen_cfile.py for strings
containing the sequence "\x00" and replace them with chr(0) before calculating
the string size.

Signed-off-by: Adam Duskett <adam.duskett@amarulasolutions.com>