gen_cfile: add legacy c file generation option
The Laerdal version of python-objdictgen changes the generation of c files from
the legacy objdictgen in an API-breaking manner. However, the code that
generates the .c file does not differ very much, and it is possible to easily
add an option to generate legacy .c files for users who wish to continue using
the legacy-generated c file.

- Add a new dropdown to the "Build Dictionary" popup dialog below
  "CANFestival C files (*.c)" called "CanFestival legacy C files (*.c)"

- Create a new directory, "gen_cfile."

- Move all definitions and methods in gen_cfile.py except for the
  GenerateFileContent and the GenerateFile methods to a new "common.py" file.

- Create a src/objdictgen/gen_cfile/standard.py file and rename the
  GenerateFileContent method to GenerateStandardFileContent.

- Create a src/objdictgen/gen_cfile/legacy.py file and rename the
  GenerateFileContent method to GenerateLegacyFileContent.

- Modify the GenerateLegacyFileContent method to generate exact matches of
  what the legacy python-objdictgen program produced.

- Modify __main__.py to accept standard arguments from argparse, and add the
  following options:
    -l: generate legacy file output
    -i: input file name
    -o: output file name

- Fix any failing tests from the new legacy logic.
- Add _legacy files to tests/od/legacy-compare/
- Add test_objdictgen_legacy to test_objdictgen.py

Even though the differences between the current and legacy dictionary files
are minimal, for clarity and ease of maintenance purposes, the entirety of
the GenerateFileContent is copied for both gen_cfile/standard.py and
gen_cfile/legacy.py, as no changes are expected to happen to legacy.py.